requirements.txt
PyYAML
boto3
jinja2
